validate: plus script all validation added

+++ controllers/plusscript.controller.js
@@ -1,16 +1,29 @@
 const plusscriptService = require("../services/plusscript.service");
-
+const Joi = require("joi");
 class plusscriptController {
   constructor() {
     this.plusscriptService = new plusscriptService();
   }
   createplusscript = async (req, res) => {
     try {
+      const contentschema = Joi.object({
+        content: Joi.string().required(),
+      });
+
       const { content } = req.body;
       const { scriptId } = req.params;
       const { userId } = res.locals.user;
+
+      let tempresult = contentschema.validate(req.body);
+      if (tempresult.error) {
+        return res.status(400).send(tempresult.error.details[0].message);
+      }
+
       if (!content) {
         return res.status(400).send("content is required.");
+      }
+      if (!scriptId) {
+        return res.status(400).send("scriptId is required.");
       }
       const plusscript = await this.plusscriptService.createplusscript({
         ScriptId: scriptId,
@@ -29,21 +42,25 @@ class plusscriptController {
       const { plusScriptId } = req.params;
       const { userId } = res.locals.user;
       if (!content) {
-        return res.status(400).send("content is required.");
+        return res.status(400).send("plus script is required.");
       }
       const existPlusScript = await this.plusscriptService.findOnescript({
         plusScriptId,
       });
+
+      if (existPlusScript == null)
+        return res.status(400).send("theres no existing plus script.");
+
       if (existPlusScript.content == content) {
         return res.status(400).send("content need to be changed.");
       }
-      if (content) {
-        await this.plusscriptService.modifyingPlusscript({
-          plusScriptId,
-          UserId: userId,
-          content,
-        });
-      }
+
+      await this.plusscriptService.modifyingPlusscript({
+        plusScriptId,
+        UserId: userId,
+        content,
+      });
+
       const updatedResult = await this.plusscriptService.findOnescript({
         plusScriptId,
       });
@@ -60,6 +77,11 @@ class plusscriptController {
         plusScriptId,
         UserId: userId,
       });
+      if (!plusScriptId)
+        return res.status(400).send("plus script id is required.");
+      if (willdeleted == null)
+        return res.status(400).send("theres no to be deleted.");
+
       res.json({ deleted: willdeleted });
       await this.plusscriptService.deletePlusscript({
         UserId: userId,
@@ -75,6 +97,10 @@ class plusscriptController {
       const plusscript = await this.plusscriptService.findOnescript({
         plusScriptId,
       });
+      if (!plusScriptId)
+        return res.status(400).send("plus script input required.");
+      if (plusscript == null)
+        return res.status(400).send("there is no plus script");
       return res.status(200).send({ plusscript });
     } catch (err) {
       console.log(err);
@@ -86,10 +112,15 @@ class plusscriptController {
       let { page } = req.query;
       page = parseInt(page);
       // console.log(page); => 1
+      if (!page) {
+        return res.status(400).send("page input required.");
+      }
       const plusscript3s = await this.plusscriptService.getting3plusscript({
         page,
       });
-      // console.log(plusscript3s);
+      if (plusscript3s == "")
+        return res.status(400).send("no more plus script");
+      console.log(plusscript3s);
       return res.send({ plusscript3s });
     } catch (err) {
       console.log(err);
